--- a/pkg/sync/expand/cycle.go
+++ b/pkg/sync/expand/cycle.go
@@ -3,114 +3,68 @@ package expand
 import (
 	"context"
 
+	"github.com/conductorone/baton-sdk/pkg/sync/expand/scc"
 	mapset "github.com/deckarep/golang-set/v2"
 )
-
-const (
-	colorWhite uint8 = iota
-	colorGray
-	colorBlack
-)
-
-// cycleDetector encapsulates coloring state for cycle detection on an
-// EntitlementGraph. Node IDs are dense (1..NextNodeID), so slices are used for
-// O(1) access and zero per-op allocations.
-type cycleDetector struct {
-	g      *EntitlementGraph
-	state  []uint8
-	parent []int
-}
-
-func newCycleDetector(g *EntitlementGraph) *cycleDetector {
-	cd := &cycleDetector{
-		g:      g,
-		state:  make([]uint8, g.NextNodeID+1),
-		parent: make([]int, g.NextNodeID+1),
-	}
-	for i := range cd.parent {
-		cd.parent[i] = -1
-	}
-	return cd
-}
-
-// dfs performs a coloring-based DFS from u, returning the first detected cycle
-// as a slice of node IDs or nil if no cycle is reachable from u.
-func (cd *cycleDetector) dfs(u int) ([]int, bool) {
-	// Self-loop fast path.
-	if nbrs, ok := cd.g.SourcesToDestinations[u]; ok {
-		if _, ok := nbrs[u]; ok {
-			return []int{u}, true
-		}
-	}
-
-	cd.state[u] = colorGray
-	if nbrs, ok := cd.g.SourcesToDestinations[u]; ok {
-		for v := range nbrs {
-			switch cd.state[v] {
-			case colorWhite:
-				cd.parent[v] = u
-				if cyc, ok := cd.dfs(v); ok {
-					return cyc, true
-				}
-			case colorGray:
-				// Back-edge to a node on the current recursion stack.
-				// Reconstruct cycle by walking parents from u back to v (inclusive), then reverse.
-				cycle := make([]int, 0, 8)
-				for x := u; ; x = cd.parent[x] {
-					cycle = append(cycle, x)
-					if x == v || cd.parent[x] == -1 {
-						break
-					}
-				}
-				for i, j := 0, len(cycle)-1; i < j; i, j = i+1, j-1 {
-					cycle[i], cycle[j] = cycle[j], cycle[i]
-				}
-				return cycle, true
-			}
-		}
-	}
-	cd.state[u] = colorBlack
-	return nil, false
-}
-
-// FindAny scans all nodes and returns the first detected cycle or nil if none exist.
-func (cd *cycleDetector) FindAny() []int {
-	for nodeID := range cd.g.Nodes {
-		if cd.state[nodeID] != colorWhite {
-			continue
-		}
-		if cyc, ok := cd.dfs(nodeID); ok {
-			return cyc
-		}
-	}
-	return nil
-}
-
-// FindFrom starts cycle detection from a specific node and returns the first
-// cycle reachable from that node, or nil,false if none.
-func (cd *cycleDetector) FindFrom(start int) ([]int, bool) {
-	return cd.dfs(start)
-}
 
 // GetFirstCycle given an entitlements graph, return a cycle by node ID if it
 // exists. Returns nil if no cycle exists. If there is a single
 // node pointing to itself, that will count as a cycle.
-func (g *EntitlementGraph) GetFirstCycle() []int {
+func (g *EntitlementGraph) GetFirstCycle(ctx context.Context) []int {
 	if g.HasNoCycles {
 		return nil
 	}
-	cd := newCycleDetector(g)
-	return cd.FindAny()
+	comps := g.ComputeCyclicComponents(ctx)
+	if len(comps) == 0 {
+		return nil
+	}
+	return comps[0]
+}
+
+// HasCycles returns true if the graph contains any cycle.
+func (g *EntitlementGraph) HasCycles(ctx context.Context) bool {
+	if g.HasNoCycles {
+		return false
+	}
+	return len(g.ComputeCyclicComponents(ctx)) > 0
 }
 
 func (g *EntitlementGraph) cycleDetectionHelper(
 	nodeID int,
 ) ([]int, bool) {
-	// Thin wrapper around the coloring-based DFS, starting from a specific node.
-	// The provided visited/currentCycle are ignored here; coloring provides the
-	// necessary state for correctness and performance.
-	cd := newCycleDetector(g)
-	return cd.FindFrom(nodeID)
+	reach := g.reachableFrom(nodeID)
+	if len(reach) == 0 {
+		return nil, false
+	}
+	adj := g.toAdjacency(reach)
+	groups := scc.CondenseFWBWGroupsFromAdj(context.Background(), adj, scc.DefaultOptions())
+	for _, comp := range groups {
+		if len(comp) > 1 || (len(comp) == 1 && adj[comp[0]][comp[0]] != 0) {
+			return comp, true
+		}
+	}
+	return nil, false
+}
+
+func (g *EntitlementGraph) FixCycles(ctx context.Context) error {
+	return g.FixCyclesFromComponents(ctx, g.ComputeCyclicComponents(ctx))
+}
+
+// ComputeCyclicComponents runs SCC once and returns only cyclic components.
+// A component is cyclic if len>1 or a singleton with a self-loop.
+func (g *EntitlementGraph) ComputeCyclicComponents(ctx context.Context) [][]int {
+	if g.HasNoCycles {
+		return nil
+	}
+	adj := g.toAdjacency(nil)
+	groups := scc.CondenseFWBWGroupsFromAdj(ctx, adj, scc.DefaultOptions())
+	cyclic := make([][]int, 0)
+	for _, comp := range groups {
+		if len(comp) > 1 || (len(comp) == 1 && adj[comp[0]][comp[0]] != 0) {
+			cyclic = append(cyclic, comp)
+		}
+	}
+	return cyclic
 }
 
 // removeNode obliterates a node and all incoming/outgoing edges.
@@ -147,34 +101,33 @@ func (g *EntitlementGraph) removeNode(nodeID int) {
 	delete(g.SourcesToDestinations, nodeID)
 }
 
-// FixCycles if any cycles of nodes exist, merge all nodes in that cycle into a
-// single node and then repeat. Iteration ends when there are no more cycles.
-func (g *EntitlementGraph) FixCycles(ctx context.Context) error {
+// FixCyclesFromComponents merges all provided cyclic components in one pass.
+func (g *EntitlementGraph) FixCyclesFromComponents(ctx context.Context, cyclic [][]int) error {
 	if g.HasNoCycles {
 		return nil
 	}
-	for {
+	if len(cyclic) == 0 {
+		g.HasNoCycles = true
+		return nil
+	}
+	for _, comp := range cyclic {
 		select {
 		case <-ctx.Done():
 			return ctx.Err()
 		default:
 		}
-		cycle := g.GetFirstCycle()
-		if cycle == nil {
-			g.HasNoCycles = true
-			return nil
-		}
-
-		if err := g.fixCycle(cycle); err != nil {
+		if err := g.fixCycle(comp); err != nil {
 			return err
 		}
 	}
+	g.HasNoCycles = true
+	return nil
 }
 
 // fixCycle takes a list of Node IDs that form a cycle and merges them into a
 // single, new node.
 func (g *EntitlementGraph) fixCycle(nodeIDs []int) error {
-	entitlementIDs := mapset.NewSet[string]()
+	entitlementIDs := mapset.NewThreadUnsafeSet[string]()
 	outgoingEdgesToResourceTypeIDs := map[int]mapset.Set[string]{}
 	incomingEdgesToResourceTypeIDs := map[int]mapset.Set[string]{}
 	for _, nodeID := range nodeIDs {
@@ -190,7 +143,7 @@ func (g *EntitlementGraph) fixCycle(nodeIDs []int) error {
 					if edge, ok := g.Edges[edgeID]; ok {
 						resourceTypeIDs, ok := incomingEdgesToResourceTypeIDs[sourceNodeID]
 						if !ok {
-							resourceTypeIDs = mapset.NewSet[string]()
+							resourceTypeIDs = mapset.NewThreadUnsafeSet[string]()
 						}
 						for _, resourceTypeID := range edge.ResourceTypeIDs {
 							resourceTypeIDs.Add(resourceTypeID)
@@ -206,7 +159,7 @@ func (g *EntitlementGraph) fixCycle(nodeIDs []int) error {
 					if edge, ok := g.Edges[edgeID]; ok {
 						resourceTypeIDs, ok := outgoingEdgesToResourceTypeIDs[destinationNodeID]
 						if !ok {
-							resourceTypeIDs = mapset.NewSet[string]()
+							resourceTypeIDs = mapset.NewThreadUnsafeSet[string]()
 						}
 						for _, resourceTypeID := range edge.ResourceTypeIDs {
 							resourceTypeIDs.Add(resourceTypeID)