… pkg/sync/expand/cycle_benchmark_test.go
… pkg/sync/expand/graph_test.go
… pkg/sync/expand/large_scc_benchmark_test.go
… pkg/sync/expand/scc/scc_test.go
… pkg/sync/syncer.go