```
benchmark                                           old ns/op        new ns/op     delta
BenchmarkSCC_FixCycles_Large/N=10000_M=100000-8     29685173578      304004762     -98.98%
BenchmarkSCC_FixCycles_Large/N=20000_M=200000-8     137692330860     702017885     -99.49%

benchmark                                           old allocs     new allocs     delta
BenchmarkSCC_FixCycles_Large/N=10000_M=100000-8     44297909       175621         -99.60%
BenchmarkSCC_FixCycles_Large/N=20000_M=200000-8     176603238      350899         -99.80%

benchmark                                           old bytes       new bytes     delta
BenchmarkSCC_FixCycles_Large/N=10000_M=100000-8     3627974760      53200072      -98.53%
BenchmarkSCC_FixCycles_Large/N=20000_M=200000-8     14439696344     106345592     -99.26%
```
<!-- This is an auto-generated comment: release notes by coderabbit.ai -->
## Summary by CodeRabbit

- **New Features**
  - Cycle detection now uses SCC-based component analysis and supports context-aware, cancellable operations for more responsive syncs.

- **Refactor**
  - Cycle handling rewritten to iterate over computed components (instead of recursive/coloring), improving scalability, determinism, and concurrency friendliness.

- **Tests**
  - Tests and benchmarks updated for context-aware APIs, deterministic SCC validation, and added large-scale SCC benchmarks.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->